requirements.txt
streamlit
langchain
google-generativeai
python-docx
pdfplumber
pandas 